DataSourceUItilities: use top-level namespaces and break out comparer classes
